updated multilive datagrids for sorting & grouping; added datagram to EFC import conversion - need to work on when data such as userIDs are null and are entity keys

+++ StreamerBotLib/DataSQL/Models/Quotes.cs
@@ -4,9 +4,9 @@ namespace StreamerBotLib.DataSQL.Models
 {
     [PrimaryKey(nameof(Number))]
     [Index(nameof(Number))]
-    public class Quotes(short number = 0, string quote = null) : EntityBase
+    public class Quotes(int number = 0, string quote = null) : EntityBase
     {
-        public short Number { get; set; } = number;
+        public int Number { get; set; } = number;
         public string Quote { get; set; } = quote;
     }
 }
